Initial packing / unpacking support
