+++ src/cluster_copier.cc
@@ -1,6 +1,7 @@
 //---------------------------------------------------------------------
 //---------------------------------------------------------------------
 #include <cluster_copier.h>
+#include <lv_message.h>
 
 //---------------------------------------------------------------------
 //---------------------------------------------------------------------
