+++ src/cluster_copier.h
@@ -1,11 +1,17 @@
 //---------------------------------------------------------------------
 //---------------------------------------------------------------------
+#pragma once
+
+//---------------------------------------------------------------------
+//---------------------------------------------------------------------
 #include <grpc_server.h>
 #include <lv_interop.h>
 #include <memory>
 #include <string>
 #include <map>
 #include <mutex>
+#include <message_metadata.h>
+#include <message_value.h>
 
 //---------------------------------------------------------------------
 //---------------------------------------------------------------------
