+++ src/event_data.cc
@@ -1,6 +1,7 @@
 //---------------------------------------------------------------------
 //---------------------------------------------------------------------
 #include <grpc_server.h>
+#include <lv_message.h>
 
 //---------------------------------------------------------------------
 //---------------------------------------------------------------------
