+++ src/lv_interop.h
@@ -2,11 +2,15 @@
 //---------------------------------------------------------------------
 #pragma once
 
+//---------------------------------------------------------------------
+//---------------------------------------------------------------------
 #ifdef _WIN32
 #define WIN32_LEAN_AND_MEAN
 #include <windows.h>
 #endif
 
+//---------------------------------------------------------------------
+//---------------------------------------------------------------------
 #include <string>
 
 //---------------------------------------------------------------------
