+++ src/serialization_session.h
@@ -1,0 +1,15 @@
+//---------------------------------------------------------------------
+// LabVIEW implementation of a gRPC Server
+//---------------------------------------------------------------------
+#pragma once
+
+//---------------------------------------------------------------------
+//---------------------------------------------------------------------
+#include <message_metadata.h>
+
+//---------------------------------------------------------------------
+//---------------------------------------------------------------------
+class LabVIEWSerializationSession : public MessageElementMetadataOwner
+{
+};
+
